+++ src/main/java/com/koc/user/application/service/UserService.java
@@ -1,7 +1,6 @@
 package com.koc.user.application.service;
 
 import com.koc.user.application.dto.UserDto;
-import com.koc.user.application.service.domainService.UserDomainService;
 import com.koc.user.domain.User;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
@@ -10,17 +9,17 @@ import org.springframework.stereotype.Service;
 @Service
 @RequiredArgsConstructor
 public class UserService {
-    private final UserDomainService userDomainService;
+    private final com.koc.user.domain.UserService userService;
 
     public User withdraw(Long id) {
-        User user = userDomainService
+        User user = userService
                 .findById(id)
                 .orElseThrow(() -> new IllegalArgumentException("해당 유저가 존재 하지 않습니다."));
         user.withdraw();
-        return userDomainService.save(user);
+        return userService.save(user);
     }
 
     public UserDto findById(long id) {
-        return userDomainService.findById(id).orElseThrow(() -> new IllegalArgumentException("없는 사용자")).toDto();
+        return userService.findById(id).orElseThrow(() -> new IllegalArgumentException("없는 사용자")).toDto();
     }
 }
