+++ src/main/java/com/koc/user/domain/UserService.java
@@ -1,7 +1,6 @@
-package com.koc.user.application.service.domainService;
+package com.koc.user.domain;
 
 import com.koc.user.application.repository.UserRepository;
-import com.koc.user.domain.User;
 import com.koc.user.infrastructure.Entity.UserEntity;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
@@ -10,7 +9,7 @@ import java.util.Optional;
 
 @Service
 @RequiredArgsConstructor
-public class UserDomainService {
+public class UserService {
     private final UserRepository userRepository;
 
     public Optional<User> findByKakaoId(Long Id) {
